*REFACTORED* ClientService and CourierService to use Spring's @Service annotation, updated method signatures for courier registration and delivery, modified JdbiClientRepository and JdbiCourierRepository for improved query handling, and adjusted Location class to allow nullable address. Project finally clean builds.

+++ code/backend/LiftDropRepository/src/main/kotlin/liftdrop/repository/RequestRepository.kt
@@ -8,8 +8,8 @@ interface RequestRepository {
     ): Int
 
     fun updateRequest(
-        requestId: Long,
-        courierId: Long?,
+        requestId: Int,
+        courierId: Int?,
         requestStatus: String?,
         eta: String?,
     ): Boolean
